image/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
cvzone==1.6.1
distlib==0.3.9
Django==5.1.6
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.1
MarkupSafe==3.0.2
MouseInfo==0.1.3
mymongo==0.1.1
mysql-connector==2.2.9
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.2.3
pdfkit==1.0.0
pillow==11.0.0
platformdirs==4.3.6
plyer==2.1.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==1.10.21
PyGetWindow==0.0.9
pyjokes==0.8.3
pymongo==4.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-engineio==4.11.2
python-socketio==5.12.1
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==308
qrcode==8.0
querytyper==0.1.0
requests==2.32.3
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.12.2
tzdata==2024.2
ulid-py==1.1.0
urllib3==2.3.0
virtualenv==20.29.1
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
WTForms==3.2.1
